mkw-dashboard-api/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
psycopg2-binary>=2.9.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.26.0

# Validation and serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
